put services in GeoConcerns namespace
