+++ app/presenters/geo_concerns/raster_work_show_presenter.rb
@@ -1,6 +1,6 @@
 module GeoConcerns
   class RasterWorkShowPresenter < GeoConcernsShowPresenter
     self.work_presenter_class = ::GeoConcerns::VectorWorkShowPresenter
-    self.file_format_service = RasterFormatService
+    self.file_format_service = GeoConcerns::RasterFormatService
   end
 end
