+++ app/services/geo_concerns/image_format_service.rb
@@ -1,0 +1,6 @@
+module GeoConcerns
+  class ImageFormatService
+    include GeoConcerns::AuthorityService
+    self.authority = Qa::Authorities::Local.subauthority_for('image_formats')
+  end
+end
