+++ app/services/geo_concerns/metadata_format_service.rb
@@ -1,0 +1,6 @@
+module GeoConcerns
+  class MetadataFormatService
+    include GeoConcerns::AuthorityService
+    self.authority = Qa::Authorities::Local.subauthority_for('metadata_formats')
+  end
+end
