+++ app/services/geo_concerns/raster_format_service.rb
@@ -1,0 +1,6 @@
+module GeoConcerns
+  class RasterFormatService
+    include GeoConcerns::AuthorityService
+    self.authority = Qa::Authorities::Local.subauthority_for('raster_formats')
+  end
+end
