+++ app/services/geo_concerns/vector_format_service.rb
@@ -1,0 +1,6 @@
+module GeoConcerns
+  class VectorFormatService
+    include GeoConcerns::AuthorityService
+    self.authority = Qa::Authorities::Local.subauthority_for('vector_formats')
+  end
+end
