--- app/services/image_format_service.rb
@@ -1,4 +1,0 @@
-class ImageFormatService
-  include AuthorityService
-  self.authority = Qa::Authorities::Local.subauthority_for('image_formats')
-end
