--- app/services/metadata_format_service.rb
@@ -1,4 +1,0 @@
-class MetadataFormatService
-  include AuthorityService
-  self.authority = Qa::Authorities::Local.subauthority_for('metadata_formats')
-end
