--- app/services/raster_format_service.rb
@@ -1,4 +1,0 @@
-class RasterFormatService
-  include AuthorityService
-  self.authority = Qa::Authorities::Local.subauthority_for('raster_formats')
-end
