--- app/services/vector_format_service.rb
@@ -1,4 +1,0 @@
-class VectorFormatService
-  include AuthorityService
-  self.authority = Qa::Authorities::Local.subauthority_for('vector_formats')
-end
